Eliminate code generation for Struct methods
Struct.to_binary, Struct.parse_value and Struct.parse_binary were doing
code generation to produce and store a specialised method for each
struct instance. Apparently (see #12) this was a necessary optimisation
in the distant past, but it's almost certainly no longer needed, and it
makes the code considerably harder to follow.

On my machine the tests run slightly but noticeably faster with this
change, but of course the tests are probably not a realistic sample for
performance testing.

This doesn't directly impact Python 3 support, but it simplifies things.