Fix encode/decode errors in structs.
The default encoding in Python 3 is 'utf-8', which is not a good choice
for byte-level data pack/unpack because it uses multiple bytes for code
point >= 128.

This commit changes encoding being used to 'latin-1', which covers all
possible code points in range (0-255) with a simple one-to-one mapping.

This commit has fixed some *UnicodeDecodeError* in practice.