Use a larger recv buffer.
The `send_and_recv` function is very slow on receiving a large volume of
data because it only receives 2048 bytes at a time and keeps appending
these small chunks of data into a large array.

This commit sets the recv buffer to a larger value. This has provable
performance benefits: The running time of a Xlib program has been
reduced from 12 seconds to 0.5 seconds, a 24x speedup.